+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # main
 
+* Added cop `Added Grape/PreferNamespace`
+
 # v0.1.0
 
 * First version